# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API Endpoint for handling organizational financial reimbursement requests",
        "title": "Financial Request Handler API",
        "contact": {
            "name": "APP Support",
            "email": "app_support@norainc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/petty_cash": {
            "put": {
                "description": "allows a logged in user to edit their pending or reject petty cash requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "petty cash",
                    "no-cache"
                ],
                "summary": "edit petty cash request",
                "operationId": "edit-petty_cash",
                "parameters": [
                    {
                        "description": "petty cash request information to update",
                        "name": "petty-cash-request-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EditPettyCash"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PettyCashOverviewRes"
                        }
                    }
                }
            },
            "post": {
                "description": "creates a petty cash request for a logged in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "petty cash",
                    "no-cache"
                ],
                "summary": "add petty cash request",
                "operationId": "add-petty_cash",
                "parameters": [
                    {
                        "description": "new petty cash request information",
                        "name": "petty-cash-request-info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PettyCashInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.PettyCashOverviewRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "allows a logged in user to delete and removes one of their petty cash requests",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "petty cash",
                    "no-cache"
                ],
                "summary": "delete petty cash request",
                "operationId": "delete-petty_cash",
                "parameters": [
                    {
                        "description": "petty cash request id to delete",
                        "name": "petty-cash-request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindPettyCashInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PettyCashOverviewRes"
                        }
                    }
                }
            }
        },
        "/petty_cash/approve": {
            "post": {
                "description": "allows an administrative user to approve a specific petty cash request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "petty cash",
                    "no-cache",
                    "admin"
                ],
                "summary": "approve petty cash request",
                "operationId": "approve-petty_cash",
                "parameters": [
                    {
                        "description": "petty cash request id to approve",
                        "name": "request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApproveRejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PettyCashOverviewRes"
                        }
                    }
                }
            }
        },
        "/petty_cash/detail": {
            "get": {
                "description": "generates detailed information for a specific petty cash request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "petty cash"
                ],
                "summary": "petty cash request detail",
                "operationId": "petty_cash-detail",
                "parameters": [
                    {
                        "description": "petty cash request id to find",
                        "name": "petty-cash-request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.FindPettyCashInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PettyCashRes"
                        }
                    }
                }
            }
        },
        "/petty_cash/monthly": {
            "get": {
                "description": "creates a monthly petty cash report for the entire organization for a given month and year",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "petty cash",
                    "reports",
                    "admin"
                ],
                "summary": "monthly petty cash request report",
                "operationId": "monthly-petty_cash",
                "parameters": [
                    {
                        "description": "month and year for report on organization wide petty cash requests",
                        "name": "month-year-input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MonthlyRequestInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PettyCashOverviewsRes"
                        }
                    }
                }
            }
        },
        "/petty_cash/reject": {
            "post": {
                "description": "allows a logged in administrative user to reject a petty cash request until further edits have been made",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "petty cash",
                    "no-cache",
                    "admin"
                ],
                "summary": "reject petty cash request",
                "operationId": "reject-petty_cash",
                "parameters": [
                    {
                        "description": "petty cash request id to reject",
                        "name": "request-id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ApproveRejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.PettyCashOverviewRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Action": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "models.ApproveRejectRequest": {
            "type": "object",
            "required": [
                "request_id"
            ],
            "properties": {
                "request_id": {
                    "type": "string"
                }
            }
        },
        "models.Category": {
            "type": "string",
            "enum": [
                "IOP",
                "INTAKE",
                "PEERS",
                "ACT_TEAM",
                "IHBT",
                "PERKINS",
                "MENS_HOUSE",
                "NEXT_STEP",
                "LORAIN",
                "PREVENTION",
                "ADMINISTRATIVE",
                "FINANCE"
            ],
            "x-enum-varnames": [
                "IOP",
                "INTAKE",
                "PEERS",
                "ACT_TEAM",
                "IHBT",
                "PERKINS",
                "MENS_HOUSE",
                "NEXT_STEP",
                "LORAIN",
                "PREVENTION",
                "ADMINISTRATIVE",
                "FINANCE"
            ]
        },
...
        "models.EditPettyCash": {
            "type": "object",
            "required": [
                "amount",
                "category",
                "date",
                "description",
                "grant_id",
                "id",
                "receipts",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "receipts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "models.FindPettyCashInput": {
            "type": "object",
            "required": [
                "petty_cash_id"
            ],
            "properties": {
                "petty_cash_id": {
                    "type": "string"
                }
            }
        },
...
        "models.Month": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12
            ],
            "x-enum-varnames": [
                "January",
                "February",
                "March",
                "April",
                "May",
                "June",
                "July",
                "August",
                "September",
                "October",
                "November",
                "December"
            ]
        },
        "models.MonthlyRequestInput": {
            "type": "object",
            "required": [
                "month",
                "year"
            ],
            "properties": {
                "month": {
                    "$ref": "#/definitions/models.Month"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.PettyCashInput": {
            "type": "object",
            "required": [
                "amount",
                "category",
                "date",
                "description",
                "grant_id",
                "receipts"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "string"
                },
                "receipts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Petty_Cash_Overview": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "current_status": {
                    "type": "string"
                },
                "current_user": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Petty_Cash_Request": {
            "type": "object",
            "properties": {
                "action_history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Action"
                    }
                },
                "amount": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/models.Category"
                },
                "created_at": {
                    "type": "string"
                },
                "current_status": {
                    "type": "string"
                },
                "current_user": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "grant_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_user_before_reject": {
                    "type": "string"
                },
                "receipts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "responses.PettyCashOverviewRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Petty_Cash_Overview"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "responses.PettyCashOverviewsRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Petty_Cash_Overview"
                    }
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
        "responses.PettyCashRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Petty_Cash_Request"
                },
                "status": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
...
    }
}